Extract header validation logic
